restructure Sensei. Rename classes to includes. Move the inc folder contents  into includes/lib. Change calls to included class to point to the new locations.
