+++ includes/class-woothemes-sensei-analysis-course.php
