+++ includes/class-woothemes-sensei-analysis-lesson.php
