+++ includes/class-woothemes-sensei-analysis-user-profile.php
