+++ includes/class-woothemes-sensei-analysis.php
