+++ includes/class-woothemes-sensei-course.php
