+++ includes/class-woothemes-sensei-emails.php
