+++ includes/class-woothemes-sensei-frontend.php
