+++ includes/class-woothemes-sensei-lesson.php
