+++ includes/class-woothemes-sensei-list-table.php
