+++ includes/class-woothemes-sensei-notices.php
