+++ includes/emails/class-woothemes-sensei-email-learner-completed-course.php
