+++ includes/emails/class-woothemes-sensei-email-teacher-completed-course.php
