+++ includes/emails/class-woothemes-sensei-email-teacher-new-message.php
