+++ includes/emails/class-woothemes-sensei-teacher-new-course-assignment.php
