+++ includes/lib/woo-functions.php
@@ -1,18 +1,4 @@
 <?php
-/**
- * Functions used by plugins
- */
-if ( ! class_exists( 'WC_Dependencies' ) )
-	require_once 'class-wc-dependencies.php';
-
-/**
- * WC Detection
- */
-if ( ! function_exists( 'is_woocommerce_active' ) ) {
-	function is_woocommerce_active() {
-		return WC_Dependencies::woocommerce_active_check();
-	}
-}
 
 /**
  * Queue updates for the WooUpdater
