+++ includes/woocommerce/class-wc-dependencies.php
@@ -26,4 +26,20 @@ class WC_Dependencies {
 
 }
 
+/**
+ * Functions used by plugins
+ */
+if ( ! class_exists( 'WC_Dependencies' ) )
+    require_once 'class-wc-dependencies.php';
+
+/**
+ * WC Detection
+ */
+if ( ! function_exists( 'is_woocommerce_active' ) ) {
+    function is_woocommerce_active() {
+        return WC_Dependencies::woocommerce_active_check();
+    }
+}
+
+
