+++ woothemes-sensei.php
@@ -30,12 +30,12 @@ Domain path: /lang/
 
 	if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
-    require_once( 'classes/class-woothemes-sensei.php' );
-    require_once( 'inc/woo-functions.php' );
-    require_once( 'inc/woothemes-sensei-functions.php' );
+    require_once( 'includes/class-woothemes-sensei.php' );
+    require_once( 'includes/lib/woo-functions.php' );
+    require_once( 'includes/lib/woothemes-sensei-functions.php' );
 
     if ( ! is_admin() ) {
-        require_once( 'inc/woothemes-sensei-template.php' );
+        require_once( 'includes/lib/woothemes-sensei-template.php' );
     }
 
     /**
